feat(my-kitchen): Phase 1 PR5 — grocery move into hub
