+++ src/components/UserSidePanel.svelte
@@ -320,7 +320,7 @@
           </li>
           <li>
             <button
-              on:click={() => navigate('/grocery')}
+              on:click={() => navigate('/my-kitchen/grocery')}
               class="w-full flex items-center gap-4 px-4 py-3 rounded-xl hover:bg-opacity-50 transition-colors cursor-pointer"
               style="color: var(--color-text-primary);"
             >
